Refactor: dto 패키지 생성
domain과 dto를 제대로 알지 못해 혼용하고 있어 이를 학습하고 패키지를 분리했습니다.
