+++ src/main/java/kr/codesqaud/cafe/controller/UserController.java
@@ -1,7 +1,7 @@
 package kr.codesqaud.cafe.controller;
 
 import kr.codesqaud.cafe.domain.User;
-import kr.codesqaud.cafe.domain.UserForm;
+import kr.codesqaud.cafe.dto.UserForm;
 import kr.codesqaud.cafe.service.UserService;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
