+++ src/main/java/kr/codesqaud/cafe/dto/LoginForm.java
@@ -1,4 +1,4 @@
-package kr.codesqaud.cafe.domain;
+package kr.codesqaud.cafe.dto;
 
 public class LoginForm {
     private String userId;
