+++ src/main/java/kr/codesqaud/cafe/service/ArticleService.java
@@ -1,10 +1,9 @@
 package kr.codesqaud.cafe.service;
 
 import kr.codesqaud.cafe.domain.Article;
-import kr.codesqaud.cafe.domain.ArticleForm;
+import kr.codesqaud.cafe.dto.ArticleForm;
 import kr.codesqaud.cafe.repository.ArticleRepository;
 import kr.codesqaud.cafe.repository.JdbcArticleRepository;
-import kr.codesqaud.cafe.repository.MemoryArticleRepository;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
